feat(*): updated comments and changed signatures of get, get-many and get-keys
This commit has changed many function signatures to account for the scenario
where the key does not exist. Instead of returning an error for non-existance keys,
the expected behvaior should be returning a none for the client to handle.

The other changes are for providing more clarity to the behavior of the APIs
ine `wasi:keyvalue` interfaces. It explained what it means to have batch operations
without guarantees to atomicity and what are atomic operations like CAS.

It also changed the portability criteria to follow the newest WASI proposal template,
and added a new goal regarding keyvalue operation performance in this proposal.

Signed-off-by: Jiaxiao Zhou (Mossaka) <duibao55328@gmail.com>